fixed bug in chol gpu kernel

+++ tests/drivers/factor_gpu_test.cxx
@@ -21,8 +21,8 @@ int main(int argc, char** argv) {
 
    // int m = 4;
    // chol_test<double>(m);
-   // chol_test<double>(opts.m);
-   chol_test<float>(opts.m);
+   chol_test<double>(opts.m);
+   // chol_test<float>(opts.m);
    
    return ret;
 }
